[stateless] move aggregate check function to plan rewriter
