+++ tests/suites/0_stateless/03_0003_select_group_by.result
@@ -55,3 +55,8 @@ SELECT number%3 as c1 FROM numbers_mt(10) where number > 2 group by c1 order by 
 |    1 |
 |    2 |
 +------+
+--------------
+SELECT number%3 as c1, number as c2 FROM numbers_mt(10) where number > 2 group by c1 order by c1
+--------------
+
+ERROR 1105 (HY000) at line 9: Code: 26, displayText = Column `number as c2` is not under aggregate function and not in GROUP BY.
