+++ tests/suites/0_stateless/03_0003_select_group_by.sql
@@ -5,3 +5,5 @@ SELECT number%3 as c1, number%2 as c2 FROM numbers_mt(10000) where number > 2 gr
 
 SELECT number%3 as c1 FROM numbers_mt(10) where number > 2 group by number order by c1;
 SELECT number%3 as c1 FROM numbers_mt(10) where number > 2 group by c1 order by c1;
+
+SELECT number%3 as c1, number as c2 FROM numbers_mt(10) where number > 2 group by c1 order by c1;
